Revert "Switch back to float, fix all unit tests"
This reverts commit 6a762092438b258ba840ebad0352fecd251e7e58.

+++ src/include/rplustreedisk/node.h
@@ -63,7 +63,7 @@ namespace rplustreedisk
     struct Partition
     {
         unsigned dimension;
-        float location;
+        double location;
     };
